refactor(#3500): split the hashable projection out of value.rs (campsite ratchet)
value.rs was 730 on origin/main and 1034 after the merge, so #3500's change is
what pushes it over the ~800 target — the split is ours, not #3504's.

Seam is a responsibility boundary: value.rs keeps the ORDINARY host conversion
(value_to_py and its *_to_py helpers, the Udt pyclass, build_udt, Duration,
inet, register_value), and the new value_hashable.rs owns the HASHABLE
PROJECTION (value_to_hashable_key, json_to_hashable_key, contains_udt).

Preserved through the move:
- all three #[deny(clippy::wildcard_enum_match_arm)] attributes travel WITH
  their functions, so AC3's mechanical pin still applies (RED-verified next);
- contains_udt and value_to_hashable_key stay in the SAME module, with the
  matched-pair comment intact — splitting them would undo round 1's finding;
- the single-site Value::Json reachability statement moves as one block.

Widened to pub(crate): build_udt and json_to_py (called across the seam). No
public surface change — both modules are private in lib.rs.
